Fixes to Teraslice Generator & Update Dependencies (#938)
* fix es version in jsconfig.json

* update dependencies

* update generator dependencies

* fixes to generator teraslice

* disable noUnusedParameters and enable forceConsistentCasingInFileNames in tsconfig

* add better debug support to travis

* use debug logger when needed

* improved debug usages in travis

* Fixes and improved logging

* fixes to travis and execution-controller/index.js

* fixes to logging in teraslice-messaging and use 4 workers for tests not 6

* fixes to build and logging

* downgrade convict

* remove some unneeded changes for this branch

+++ packages/generator-teraslice/generators/package/index.js
@@ -97,35 +97,27 @@ module.exports = class extends Generator {
                 this.templatePath('tsconfig.build.json'),
                 this.destinationPath('tsconfig.build.json')
             );
+        }
 
+        const folderName = this.props.typescript ? 'src' : 'lib';
+        const ext = this.props.typescript ? '.ts' : '.js';
+
+        if (!this.fs.exists(this.destinationPath(folderName))) {
             this.fs.copy(
-                this.templatePath('src'),
-                this.destinationPath('src')
+                this.templatePath(folderName),
+                this.destinationPath(folderName)
+            );
+        }
+
+        if (!this.fs.exists(this.destinationPath('test'))) {
+            this.fs.copy(
+                this.templatePath(`test/index-spec${ext}`),
+                this.destinationPath(`test/index-spec${ext}`)
             );
 
             this.fs.copy(
-                this.templatePath('test/index-spec.ts'),
-                this.destinationPath('test/index-spec.ts')
-            );
-
-            this.fs.copy(
-                this.templatePath('test/example-spec.ts'),
-                this.destinationPath('test/example-spec.ts')
-            );
-        } else {
-            this.fs.copy(
-                this.templatePath('lib'),
-                this.destinationPath('lib')
-            );
-
-            this.fs.copy(
-                this.templatePath('test/index-spec.js'),
-                this.destinationPath('test/index-spec.js')
-            );
-
-            this.fs.copy(
-                this.templatePath('test/example-spec.js'),
-                this.destinationPath('test/example-spec.js')
+                this.templatePath(`test/example-spec${ext}`),
+                this.destinationPath(`test/example-spec${ext}`)
             );
         }
     }
